The real2020 probe is now correctly loaded
(not 100% sure about the rotation but it should be the same as 2017 so I have to trust myself from 2017)

I import a probe that is loaded with ptypy and resampled to THIS SPECIFIC Bragg geometry. i.e.
with resolution ~13 nm:
(psize=(2*1E-2, 55*1E-6, 55*1E-6), shape=(62, 170, 170), energy=10.0, distance=1.0, theta_bragg=8.8, propagation = "farfield")